first changes to server<> and tests running again

+++ bluetoe/service.hpp
@@ -366,7 +366,7 @@ namespace bluetoe {
         template < typename CCCDIndices, std::size_t ClientCharacteristicIndex, typename ServiceUUID, typename Server, typename ... Options >
         struct generate_attribute< service_defintion_tag, CCCDIndices, ClientCharacteristicIndex, ServiceUUID, Server, Options... >
         {
-            static attribute_access_result access( attribute_access_arguments& args, std::uint16_t )
+            static attribute_access_result access( attribute_access_arguments& args, std::size_t )
             {
                 typedef typename find_by_meta_type< service_uuid_meta_type, Options... >::type uuid;
 
@@ -423,7 +423,7 @@ namespace bluetoe {
 
             typedef service_handles< service_list, included_service > handles;
 
-            static details::attribute_access_result access( attribute_access_arguments& args, std::uint16_t )
+            static details::attribute_access_result access( attribute_access_arguments& args, std::size_t )
             {
                 static constexpr std::uint8_t value[] = {
                     handles::service_attribute_handle & 0xff,
@@ -476,7 +476,7 @@ namespace bluetoe {
 
             typedef service_handles< service_list, included_service > handles;
 
-            static details::attribute_access_result access( attribute_access_arguments& args, std::uint16_t )
+            static details::attribute_access_result access( attribute_access_arguments& args, std::size_t )
             {
                 static constexpr std::uint8_t value[] = {
                     handles::service_attribute_handle & 0xff,
